// src/@types/modules/registry.d.ts
import Address, { DatabaseAddress } from './address';
import Contact from './contact';

export interface BaseRegistry {
  name: string;
  type: string;
  cpf: string;
  uuid: string;
  cnpj?: string;
}

export default interface Registry extends Omit<BaseRegistry, 'uuid'> {
  address: Address;
  contact: Contact;
}

export interface DatabaseRegistry {
  registry: BaseRegistry;
  address: DatabaseAddress;
  contact: Contact;
}